Split Header into Frame and Header components
