+++ app/components/Frame/index.js
@@ -1,0 +1,3 @@
+import Frame from './Frame'
+
+export default Frame
