+++ app/components/Frame/styles.js
@@ -1,0 +1,9 @@
+import { StyleSheet } from 'react-native';
+import colors from '../../config/colors';
+import metrics from '../../config/metrics';
+
+export default styles = StyleSheet.create({
+	bodyContainer: {
+		flex: 1,
+	},
+});
